Add basic foundation for the new Backend abstraction
